Absorb initializing state into wallet state and simplify
